tegra: mce: update cstats to use 64 bits
NVG c-stats interface provides 64 bit data, but the current
implementation on uses 32 bits leaving out important high order
bits. This patch updates the api in tegra mce driver to pass
all 64bits through the api.

Reviewed-on: https://git-master.nvidia.com/r/1667020
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/platform/tegra/tegra-mce.c
@@ -38,7 +38,7 @@ static int tegra_mce_id = TEGRA_MCE_ID_MAX;
 int (*_tegra_mce_enter_cstate)(u32, u32);
 int (*_tegra_mce_update_cstate_info)(u32, u32, u32, u8, u32, bool);
 int (*_tegra_mce_update_crossover_time)(u32, u32);
-int (*_tegra_mce_read_cstate_stats)(u32, u32 *);
+int (*_tegra_mce_read_cstate_stats)(u32, u64 *);
 int (*_tegra_mce_write_cstate_stats)(u32, u32);
 int (*_tegra_mce_is_sc7_allowed)(u32, u32, u32 *);
 int (*_tegra_mce_online_core)(int);
@@ -116,7 +116,7 @@ EXPORT_SYMBOL_GPL(tegra_mce_update_crossover_time);
  *
  * Returns 0 if success.
  */
-int tegra_mce_read_cstate_stats(u32 state, u32 *stats)
+int tegra_mce_read_cstate_stats(u32 state, u64 *stats)
 {
 	if (!_tegra_mce_read_cstate_stats)
 		return -ENOTSUPP;
