+++ include/linux/tegra-mce.h
@@ -101,7 +101,7 @@ int tegra_mce_enter_cstate(u32 state, u32 wake_time);
 int tegra_mce_update_cstate_info(u32 cluster, u32 ccplex, u32 system,
 				 u8 force, u32 wake_mask, bool valid);
 int tegra_mce_update_crossover_time(u32 type, u32 time);
-int tegra_mce_read_cstate_stats(u32 state, u32 *stats);
+int tegra_mce_read_cstate_stats(u32 state, u64 *stats);
 int tegra_mce_write_cstate_stats(u32 state, u32 stats);
 int tegra_mce_is_sc7_allowed(u32 state, u32 wake, u32 *allowed);
 int tegra_mce_online_core(int cpu);
